requirements.txt
requests
dnspython
tqdm
fuzzywuzzy
tabulate
whois